+++ src/main/scala/org/podval/tools/scalajsplugin/GradleNames.scala
@@ -7,10 +7,13 @@ import org.gradle.api.tasks.SourceSet
 // Names to use for various Gradle things: source sets, configurations, tasks...
 final class GradleNames(suffix: String):
   private def s(string: String) = string + suffix
+  val featureName: String = s("main")
+  
   val mainSourceSetName: String = s(SourceSet.MAIN_SOURCE_SET_NAME)
   val implementationConfigurationName: String = s(JavaPlugin.IMPLEMENTATION_CONFIGURATION_NAME)
   val linkTaskName: String = s("link")
   val runTaskName: String = s("run")
+  val scalaDocTaskName: String = s("scaladoc")
 
   val testSourceSetName: String= s(SourceSet.TEST_SOURCE_SET_NAME)
   val testImplementationConfigurationName: String = s(JavaPlugin.TEST_IMPLEMENTATION_CONFIGURATION_NAME)
@@ -19,4 +22,6 @@ final class GradleNames(suffix: String):
 
   val scalaCompilerPluginsConfigurationName: String = s(ScalaBasePlugin.SCALA_COMPILER_PLUGINS_CONFIGURATION_NAME)
   val runtimeClasspathConfigurationName: String = s(JavaPlugin.RUNTIME_CLASSPATH_CONFIGURATION_NAME)
+  
+  val incrementalScalaAnalysisElementsConfigurationName: String = s("incrementalScalaAnalysisElements")
   val scalaCompileTaskName: String = s("scala")