+++ src/main/scala/org/podval/tools/scalajsplugin/nonjvm/NonJvmLinkMainTask.scala
@@ -1,10 +1,6 @@
 package org.podval.tools.scalajsplugin.nonjvm
 
-import org.gradle.api.DefaultTask
-
-abstract class NonJvmLinkMainTask[L <: NonJvmLinkTask[L]] extends DefaultTask with NonJvmLinkTask[L]:
+abstract class NonJvmLinkMainTask[L <: NonJvmLinkTask[L]] extends NonJvmLinkTask[L]:
   this: L =>
-
-  final override protected def flavourBase: String = "Link"
-
+  
   final override protected def isTest: Boolean = false