+++ src/main/scala/org/podval/tools/scalajsplugin/nonjvm/NonJvmRunTask.scala
@@ -4,9 +4,7 @@ import org.gradle.api.GradleException
 import org.podval.tools.build.Gradle
 
 trait NonJvmRunTask[L <: NonJvmLinkTask[L]] extends NonJvmTask[L]:
-  protected def linkTaskClass: Class[? <: L]
-
-  final override def flavourSuffix: String = linkTask.flavourSuffix
+  protected def linkTaskClass: Class[? <: L] // TODO assign in configureAll()
 
   final override protected def linkTask: L = Gradle
     .findDependsOnProviderOrTask(this, linkTaskClass)