+++ src/main/scala/org/podval/tools/scalajsplugin/nonjvm/NonJvmTestTask.scala
@@ -2,5 +2,4 @@ package org.podval.tools.scalajsplugin.nonjvm
 
 import org.podval.tools.test.task.TestTask
 
-abstract class NonJvmTestTask[L <: NonJvmLinkTask[L]] extends TestTask with NonJvmRunTask[L]:
-  final override protected def flavourBase: String = "Test"
+abstract class NonJvmTestTask[L <: NonJvmLinkTask[L]] extends TestTask with NonJvmRunTask[L]