+++ src/main/scala/org/podval/tools/scalajsplugin/scalanative/ScalaNativeTask.scala
@@ -2,5 +2,4 @@ package org.podval.tools.scalajsplugin.scalanative
 
 import org.podval.tools.scalajsplugin.nonjvm.NonJvmTask
 
-trait ScalaNativeTask extends NonJvmTask[ScalaNativeLinkTask]:
-  final override protected def backend: String = "Scala Native"
+trait ScalaNativeTask extends NonJvmTask[ScalaNativeLinkTask]