+++ src/main/scala/org/podval/tools/test/framework/JUnit4ScalaNative.scala
@@ -1,14 +1,13 @@
 package org.podval.tools.test.framework
 
-import org.podval.tools.build.{ScalaDependency, Version}
-import org.podval.tools.scalanative.ScalaNative
+import org.podval.tools.build.{ScalaBackendKind, ScalaDependency}
 
 object JUnit4ScalaNative extends FrameworkDescriptor(
   name = "Scala Native JUnit test framework",
   displayName = "JUnit4 Scala Native",
   group = "org.scala-native",
   artifact = "junit-runtime",
-  versionDefault = ScalaNative.versionDefault,
+  versionDefault = ScalaBackendKind.Native.versionDefault,
   className = "scala.scalanative.junit.JUnitFramework",
   sharedPackages = List("com.novocode.junit", "junit.framework", "junit.extensions", "org.junit"),
   usesTestSelectorAsNestedTestSelector = true,