+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- doc: Scala Native;
+- cleanup: use TaskProvider instead of Task more;
+- cleanup: `NonJvm`;
 
 ## [0.7.0] - 2025-05-05
 - feat: support for Scala Native;